making it a package for easier use in the future
