+++ flashstu/__init__.py
@@ -1,0 +1,3 @@
+from .model import FlashSTU, FlashSTUConfig
+from .modules.stu import STU
+from .utils.utils import get_spectral_filters
