+++ flashstu/model.py
@@ -3,10 +3,13 @@ import torch.nn as nn
 
 from transformers import PreTrainedModel
 
-from stu import STU
-from modules import Attention, MLP
-from utils import nearest_power_of_two
-from flash_stu.config import FlashSTUConfig
+from flashstu.modules.stu import STU
+from flashstu.modules.modules import Attention
+from flashstu.utils.utils import nearest_power_of_two
+from flashstu.config import FlashSTUConfig
+from flashstu.layers.stu_layer import STULayer
+from flashstu.layers.attention_layer import AttentionLayer
+
 
 try:
     from liger_kernel.transformers.swiglu import LigerSwiGLUMLP as TritonMLP
@@ -22,43 +25,6 @@ except ImportError as e:
     print(f"Unable to import Triton-based RMSNorm: {e}. Falling back to PyTorch implementation.")
     from torch.nn import RMSNorm
     triton_norm = False
-
-
-class STULayer(nn.Module):
-    def __init__(self, config, phi, n):
-        super(STULayer, self).__init__()
-        self.stu_norm = TritonNorm(config.n_embd) if triton_norm else RMSNorm(config.n_embd, dtype=config.torch_dtype)
-        self.stu = STU(config, phi, n)
-        self.mlp_norm = TritonNorm(config.n_embd) if triton_norm else RMSNorm(config.n_embd, dtype=config.torch_dtype)
-        self.mlp = TritonMLP(config) if triton_mlp else MLP(config, dtype=config.torch_dtype)
-
-        # TODO: Write Issue in Liger-Kernel repo to support user-defined dtype for MLP
-        self.stu_norm = self.stu_norm.to(dtype=config.torch_dtype)
-        self.mlp = self.mlp.to(dtype=config.torch_dtype)
-        self.mlp_norm = self.mlp_norm.to(dtype=config.torch_dtype)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        x = x + self.stu(self.stu_norm(x))
-        x = x + self.mlp(self.mlp_norm(x))
-        return x
-
-class AttentionLayer(nn.Module):
-    def __init__(self, config) -> None:
-        super(AttentionLayer, self).__init__()
-        self.attn_norm = TritonNorm(config.n_embd) if triton_norm else RMSNorm(config.n_embd, dtype=config.torch_dtype)
-        self.attn = Attention(config)
-        self.mlp_norm = TritonNorm(config.n_embd) if triton_norm else RMSNorm(config.n_embd, dtype=config.torch_dtype)
-        self.mlp = TritonMLP(config) if triton_mlp else MLP(config, dtype=config.torch_dtype)
-
-        # TODO: Write Issue in Liger-Kernel repo to support user-defined dtype for MLP
-        self.attn_norm = self.attn_norm.to(dtype=config.torch_dtype)
-        self.mlp = self.mlp.to(dtype=config.torch_dtype)
-        self.mlp_norm = self.mlp_norm.to(dtype=config.torch_dtype)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        x = x + self.attn(self.attn_norm(x))
-        x = x + self.mlp(self.mlp_norm(x))
-        return x
 
 class FlashSTU(PreTrainedModel):
     config_class = FlashSTUConfig
