--- requirements.txt
@@ -1,16 +1,0 @@
-packaging==24.1
-wheel==0.44.0
-numpy==1.26.4
-ninja==1.11.1.1
-torch==2.4.0
-triton==3.0.0
-liger-kernel==0.2.1
-einops==0.8.0
-tiktoken==0.7.0
-transformers==4.44.0
-datasets==2.20.0
-safetensors==0.4.4
-psutil==6.0.0
-matplotlib==3.9.2
-
-# Don't forget to install flash-attn and flashfftconv, per the README!
