+++ setup.cfg
@@ -1,0 +1,37 @@
+[metadata]
+name = flashstu
+version = 0.1.0
+description = An optimized, open-source PyTorch implementation of the Spectral Transform Unit (STU)
+author = Y. Isabel Liu, Windsor Nguyen
+author_email = isabel.liu@princeton.edu, minhtringuyen@princeton.edu
+license = Apache License
+license_file = LICENSE
+keywords = spectral state-space model, pytorch
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: Apache License
+    Operating System :: Unix
+
+[options]
+packages = find:
+python-requires = >=3.12
+install_requires =
+    packaging==24.1
+    wheel==0.44.0
+    numpy==1.26.4
+    ninja==1.11.1.1
+    # torch==2.4.0 You are already installing it separately
+    triton==3.0.0
+    liger-kernel==0.2.1
+    einops==0.8.0
+    tiktoken==0.7.0
+    transformers==4.44.0
+    datasets==2.20.0
+    safetensors==0.4.4
+    psutil==6.0.0
+    matplotlib==3.9.2
+
+# Don't forget to install flash-attn and flashfftconv, per the README!
+
+[options.packages.find]
+where = flashstu
